api/requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
future==0.18.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.18.4
opencv-python==4.2.0.34
pandas==1.0.3
Pillow==7.1.2
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
https://download.pytorch.org/whl/cpu/torch-1.0.1.post2-cp37-cp37m-linux_x86_64.whl
torchvision==0.2.2
Werkzeug==1.0.1
